pytest: Pass result to fixtures and keep directories of failed tests
@Reported-by: Rusty Russell <@rustyrussell>

+++ tests/fixtures.py
@@ -35,7 +35,7 @@ def test_base_dir():
 
 
 @pytest.fixture
-def directory(test_base_dir, test_name):
+def directory(request, test_base_dir, test_name):
     """Return a per-test specific directory.
 
     This makes a unique test-directory even if a test is rerun multiple times.
@@ -48,7 +48,12 @@ def directory(test_base_dir, test_name):
 
     yield directory
 
-    shutil.rmtree(directory)
+    # This uses the status set in conftest.pytest_runtest_makereport to
+    # determine whether we succeeded or failed.
+    if request.node.rep_call.outcome == 'passed':
+        shutil.rmtree(directory)
+    else:
+        logging.debug("Test execution failed, leaving the test directory {} intact.".format(directory))
 
 
 @pytest.fixture
